Removes default list root node
The old parser always worked from a list so it inserted a shadow root node. This
won't work with scheme syntax because we want to execute it as a list. So this
changes the parser to always return the first node it can parse, so both "1" and
"(+ 0 1)" should return a single 1 int node.